requirements.txt
Django ~= 2.2.10
django-mdeditor ~= 0.1.18
gunicorn ~= 20.1.0
#Markdown ~= 3.3.4
